[nullsafe] Make condition redundant slightly less aggressive
Summary:
In practice, condition redundant is extremely noisy and low-signal
warning (hence it is turned off by default).

This diff does minor tweaks, without the intention to change anything
substantially:
1/ Change severity to advice
2/ Change "is" to "might be"
3/ Describe the reason in case the origin comes from a method.

The short term motivation is to use 3/ for specific use-case: running nullsafe on codebase and
identify most suspicious functions (that are not annotated by often
compared with null).

Reviewed By: artempyanykh

Differential Revision: D19553571

fbshipit-source-id: 2b43ea0af