[selfInBlock] Remove false positive from mix self weakSelf
Summary: The translation of closures includes a load instruction for the captured variable, then we add that corresponding id to the closure. This doesn't correspond to an actual "use" of the captured variable in the source program though, and causes false positives. Here we remove the ids from the domain when that id is being added to a closure.

Reviewed By: ngorogiannis

Differential Revision: D19557352

fbshipit-source-id: 52b426011