[clang frontend] Add support for the clang attribute NS_NOESCAPE for Objective-C blocks in Sil
Summary:
This attribute is given to parameters of methods that take Objective-C blocks to show that they will be used only in the current context and won't "escape" the context.
We translate it here, with the goal to use it in a new check later. The check is about not using weakSelf in non-escaping blocks, because retain cycles are not possible.

The translation is a bit complex because the annotation comes in the parameter of a method, but in the checker we will need it in the block. So we pass it around in the frontend from the translation of the method call to the translation context and on to the block expression and the block declaration afterwards.

Reviewed By: ngorogiannis

Differential Revision: D19600377

fbshipit-source-id: dd49539bd