[SelfInBlock] Check for when weakSelf is used in a no escaping block
Summary: This adds a check for when developers use weakSelf (and maybe strongSelf) in a block, when there is no need for it, because it won't cause a retain cycle. In general there is an annotation in Objective-C for methods that take blocks, NS_NOESCAPE, that means that the passed block won't leave the scope, i.e. is "no escaping".  So we report when weakSelf is used and the block has the annotation.

Reviewed By: ngorogiannis

Differential Revision: D19662468

fbshipit-source-id: f5ac695aa