[HIL] do not crash on confusing function call expressions
Summary:
When the expression resolving to a function to be called could not be
translated to either a proc name or at least an access path, the HIL
translation code would crash. However, this is perfectly possible. Moreover, no
one actually uses the payload of the `Indirect` datatype so no complication
arises from generilising its type to `HilExp.t`, as done in this diff.

Reviewed By: ngorogiannis

Differential Revision: D19691127

fbshipit-source-id: 4c0400ab7