@allow-large-files [ocaml] upgrade core to v13
Summary:
More newer = more better.

This flips the Not_found -> Not_found_s switch, and forbids a bunch more
polymorphic comparisons (mostly turned into `int` comparisons for
convenience). Earlier diffs prepare for this so this diff is only about
breaking changes in the API, of which there are only a few.

Reviewed By: jberdine

Differential Revision: D19861583

fbshipit-source-id: fe54ce8f0