[scheduler][static callgraph] remove lazy init now that it's enforced in ProcessPool
Summary:
Building the call graph should be done only in the scheduler process after having forked all workers.  This was achieved by a lazy init pattern, whereby the first time `next` was called, it would build the call graph, on the assumption that `next` is only ever called in the scheduler after forking.

D19769741 made this compulsory regardless the scheduler by passing a thunk to `ProcessPool` which is called to obtain the actual scheduler, on the right process and after the fork.

This means we don't need the custom lazy init logic any more.  In addition, that set up used a DB query to overapproximate the number of procedures to analyse, because this was supposed to be provided *before* forking.  Now this is also not needed, and on top of that we can provide the exact number after building the call graph.

Reviewed By: ezgicicek

Differential Revision: D19833974

fbshipit-source-id: 7f6d51d93