[pulse] formal/actual length mismatch is a contradiction
Summary:
We can already tell that a summary cannot be applied by raising
`Contradiction`, so use this mechanism to stop applying a summary if the
number of formals doesn't match the number of actuals provided.
Previously we would return an option type and `None` in case of
mismatch, on top of the `raise Contradiction` mechanism (used for
aliasing and arithmetic contradictions).

This changes the behaviour of pulse in this case: before we would skip
over the function call, but now we stop the analysis.

Reviewed By: dulmarod

Differential Revision: D19940729

fbshipit-source-id: 6def40cd6