[infer] Add --continue-analysis option
Summary:
The `--continue-analysis` option enables continuing analysis after more targets are captured by
`--continue`.  For example,

```
$ infer capture -- buck build tgt1
$ infer analyze --merge
$ infer capture --continue -- bucck build tgt2
$ infer analyze --merge --continue-analyze
```

In the last analysis, it reuses the analysis results of `tgt1` from the previous analysis.  If
`tgt1` and `tgt2` have a same dependency to a library, the analysis results of the library is also
reused.

Reviewed By: dulmarod

Differential Revision: D19996598

fbshipit-source-id: bb6874a6f