[scheduler][restart] Reduce live-locking by using data produced on failure
Summary: When a worker fails because it can't a get the lock of a `Procname` it will include it in the exception that it throws so the `RestartScheduler` can record it as a dependency. Then when scheduling a new work item from `RestartScheduler.next` it will check if this dependency is already met, if it isn't it will not schedule the `Procname` yet.

Reviewed By: ngorogiannis

Differential Revision: D19820331

fbshipit-source-id: b48cacc9a