[scheduler][callgraph] load graph directly from DB
Summary:
Currently the call graph of all captured procedures is loaded and then traversed to flag reachable procedures from modified files, followed by deleting the unflagged part, and unflagging the rest.  This is a bit wasteful, and doesn't lend itself nicely to constructing directly the reverse call graph, which further diffs will do.

This diff loads all captured procedures and callees in a hashconsed table, and performs a BFS from procedures in modified files, to build the call graph in one pass.

Reviewed By: fgasperij

Differential Revision: D19888965

fbshipit-source-id: eeb59356e