[sledge] Strengthen existential witnessing for memory theory
Summary:
Strengthen computation of solution substitutions used for existential
witnessing by using the solver for the memory contents theory. This
uses a generalization of the equation solver implementation which
accepts a predicate used as a filter for equations added to the
solution substitution. When used for solving for a given set of
variables, this filter excludes equations which do not meet the
desired variable conditions.

Reviewed By: jvillard

Differential Revision: D20120275

fbshipit-source-id: 4203d5e41