[website] exclude from repo language stats
Summary:
Having infer's repo main language be HTML because of the generated OCaml
documentation is a tad sad. We could also exclude only these from the language
count but arguably the whole website is not really part of the codebase, and
there are plans to generate more of it from the code of infer itself too, which
would skew statistics even more.

Reviewed By: jberdine

Differential Revision: D20248382

fbshipit-source-id: e91dc502a