[tech debt] Remove experimental nullability checkers
Summary:
These were not used (and were actually activated byt the same config
param). They both are in experimental stage that never reached maturity.

Since the team does not have immediate plans to work on ObjC nullability
checker; and since "eradicate" (now known as nullsafe) is the main
solution for Java, removing it is sensible.

Reviewed By: jvillard

Differential Revision: D20279866

fbshipit-source-id: 79e64992b