[sledge] Move Reg.demangle to frontend
Summary:
`Reg.demangle` is implemented by calling the `_cxa_demangle` C++
runtime system function. This will be linked into the sledge binary,
due to being linked with llvm, but will not necessarily be available
in the sledge library. So make it a dynamically-set function to avoid
calling an undefined function from the library.

Reviewed By: jvillard

Differential Revision: D20323791

fbshipit-source-id: bda9afd37