[config] datatype for checkers
Summary:
The goals are to have all the checker definitions and documentation in one
place (except how to actually run them, since that's not quite the same
concept; for example inferbo is one checker but several analyses depend on its
symbolic execution), and later on to be able to link issues reported by infer
back to the checker that generated them.

This makes apparent that the documentation of our checkers is lacking,
not touching that in this diff.

Not sure if "analysis" would be a better name than "checker" at this
point? For instance "Linters" is one of the checkers, which historically
at least we have not considered to be the case.

Reviewed By: mityal

Differential Revision: D20252386

fbshipit-source-id: fc611bfb7