[debt] delete Procdesc.is_connected
Summary:
At some point we thought disconnected CFGs (where some nodes are not
reachable from the initial node) were signs of bugs in our frontend, but
it turned out not to be the case. Thus, we compute the boolean "is
connected" for each procdesc for the only purpose of logging that
uninteresting piece of information.

Delete it.

Reviewed By: ngorogiannis

Differential Revision: D20342834

fbshipit-source-id: 3f9317003