[nullsafe] Evaluate lhs nullability in assignment check based on field declaration
Summary:
When assigning to field `field = expr()` we need to compare actual
nullability of `expr()` (rhs) with formal nullability of field (lhs).

Formal nullability is based on field declaration.

Before this diff lhs nullability was calculated based on actual
nullability of lhs. This is an absolutely wrong thing to do!

E.g. `field = null` would update nullability of field in typestate, and
from this field will be considered as nullable, so next time field =
null is allowed...

Due to series of hacks pile on top of each other, this wrong behavior
actually works correctly; but this depends on completely wrong things
happening in other places of the program (e.g. typestate modifies in
wrong places to make this work!).

This diff unblocks refactoring that will clean up hacks mentioned above,
and simplify typechecking logic.

Reviewed By: artempyanykh

Differential Revision: D20334925

fbshipit-source-id: 8bbcbd33a